Cast List from Wes Anderson Films

A dataset of actors appearing in films directed by Wes Anderson. Includes film title, actor name, release year, runtime in minutes, and worldwide box office takings.

Usage

wes_actors

Format

A tibble with 5 variables:

film

Film title

actor

Actor name

release

Date of first screening

box_office_usd

Worldwide box office takings in USD

duration_period

Duration of the film in hh:mm:ss

Source

Wikidata. Query via Wikidata

Examples

# Show most frequent actors
dplyr::count(wes_actors, actor, sort = TRUE)
#> # A tibble: 353 × 3
#> # Groups:   film, actor [353]
#>    film          actor                              n
#>    <chr>         <chr>                          <int>
#>  1 Asteroid City Aaron Ziobrowski                   1
#>  2 Asteroid City Adrien Brody                       1
#>  3 Asteroid City Adrián Corpa Fernández             1
#>  4 Asteroid City Aimee Mullins                      1
#>  5 Asteroid City Ainhoa Aldanondo                   1
#>  6 Asteroid City Alejandro García Aparicio          1
#>  7 Asteroid City Alejandro Urbina Escalona          1
#>  8 Asteroid City Alicia San Lorenzo                 1
#>  9 Asteroid City Alvaro Olivas Marques              1
#> 10 Asteroid City Ana Isabel Comendador Martínez     1
#> # ℹ 343 more rows